backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-docx>=1.1.0
pydantic>=2.6.0
pyyaml>=6.0.1
python-multipart>=0.0.9
